Replace C-style bool patterns: 0!=0 \u2192 false, 1!=0 \u2192 true
624 replacements across 31 files. Fixed 2 false positives where regex
matched inside hex literal (0x1) and variable name (c2rust_fresh1).

+++ src/xkb/utils.rs
@@ -948,7 +948,7 @@ pub unsafe fn xkb_check_versioned_struct_size_(
         while p < end {
             let c2rust_fresh1 = p;
             p = p.offset(1);
-            if *c2rust_fresh1 as i32 != 0 as i32 {
+            if *c2rust_fresh1 != 0 {
                 return XKB_ERROR_ABI_FORWARD_COMPAT;
             }
         }
